moved AppTitle & Utils to their correct package
